refactor: colonizethis_orders wave 8 Slice C near-cap splits (#4317)
Split six residual near-cap lib modules (api impl, application,
development assign, work suggestion unit loop, candidacy prefilters,
standard work handler config) under the 320-line ceiling with stable
public exports.

Refs #4317
