+++ packages/colonizethis_orders/lib/src/orders/order_suggestion_api_impl.dart
@@ -1,51 +1,18 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'orders_logging.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:colonizethis_world/colonizethis_world.dart';
 
-import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'order_suggestion.dart' as suggestion;
 import 'order_suggestion_api.dart';
-import 'package:colonizethis_world/colonizethis_world.dart';
-import 'order_resolution_context.dart';
-
-List<T> _suggestWithLog<T>(String method, String playerId, List<T> Function() run) {
-  ordersLog.d('order suggestion API $method player=$playerId');
-  return run();
-}
-
-/// Bundles the four standard suggest* parameters for delegation (Refs #3500).
-class _StandardSuggestContext {
-  const _StandardSuggestContext({
-    required this.view,
-    required this.game,
-    required this.topology,
-    required this.currentOrders,
-  });
-
-  final PlayerView view;
-  final Game game;
-  final MapTopology topology;
-  final Orders currentOrders;
-
-  List<T> loggedSuggest<T>(String method, List<T> Function() invoke) =>
-      _suggestWithLog(method, view.playerId, invoke);
-}
+import 'order_suggestion_api_impl_helpers.dart';
+import 'order_suggestion_api_impl_naval_diplomatic.dart';
+import 'order_suggestion_api_impl_trade.dart';
 
 /// Default implementation of [OrderSuggestionAPI] using the top-level suggest* functions.
-class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
+class DefaultOrderSuggestionAPI
+    with OrderSuggestionAPINavalDiplomatic, OrderSuggestionAPITrade
+    implements OrderSuggestionAPI {
   const DefaultOrderSuggestionAPI();
-
-  _StandardSuggestContext _ctx(
-    PlayerView view,
-    Game game,
-    MapTopology topology,
-    Orders currentOrders,
-  ) => _StandardSuggestContext(
-    view: view,
-    game: game,
-    topology: topology,
-    currentOrders: currentOrders,
-  );
 
   @override
   List<MoveOrder> suggestMoveOrders(
@@ -54,7 +21,12 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
     MapTopology topology,
     Orders currentOrders,
   ) {
-    final ctx = _ctx(view, game, topology, currentOrders);
+    final ctx = standardSuggestContext(
+      view: view,
+      game: game,
+      topology: topology,
+      currentOrders: currentOrders,
+    );
     return ctx.loggedSuggest(
       'suggestMoveOrders turn=${game.worldState.turnState.turnNumber}',
       () => suggestion.suggestMoveOrders(
@@ -73,7 +45,12 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
     MapTopology topology,
     Orders currentOrders,
   ) {
-    final ctx = _ctx(view, game, topology, currentOrders);
+    final ctx = standardSuggestContext(
+      view: view,
+      game: game,
+      topology: topology,
+      currentOrders: currentOrders,
+    );
     return ctx.loggedSuggest(
       'suggestArmyMoveOrders',
       () => suggestion.suggestArmyMoveOrders(
@@ -93,7 +70,12 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
     Orders currentOrders, {
     Map<String, TileMapResult>? tileMapByRegion,
   }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
+    final ctx = standardSuggestContext(
+      view: view,
+      game: game,
+      topology: topology,
+      currentOrders: currentOrders,
+    );
     return ctx.loggedSuggest(
       'suggestWorkOrders',
       () => suggestion.suggestWorkOrders(
@@ -114,7 +96,12 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
     Orders currentOrders, {
     bool includeCivilianBuilds = false,
   }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
+    final ctx = standardSuggestContext(
+      view: view,
+      game: game,
+      topology: topology,
+      currentOrders: currentOrders,
+    );
     return ctx.loggedSuggest(
       'suggestBuildOrders',
       () => suggestion.suggestBuildOrders(
@@ -134,7 +121,12 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
     MapTopology topology,
     Orders currentOrders,
   ) {
-    final ctx = _ctx(view, game, topology, currentOrders);
+    final ctx = standardSuggestContext(
+      view: view,
+      game: game,
+      topology: topology,
+      currentOrders: currentOrders,
+    );
     return ctx.loggedSuggest(
       'suggestRecruitWorkerOrders',
       () => suggestion.suggestRecruitWorkerOrders(
@@ -159,7 +151,12 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
     int researchSeed = 0,
     int categoryDiversifyWeight = 0,
   }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
+    final ctx = standardSuggestContext(
+      view: view,
+      game: game,
+      topology: topology,
+      currentOrders: currentOrders,
+    );
     return ctx.loggedSuggest(
       'suggestResearchOrders',
       () => suggestion.suggestResearchOrders(
@@ -174,126 +171,6 @@ class DefaultOrderSuggestionAPI implements OrderSuggestionAPI {
         researchSeed: researchSeed,
         categoryDiversifyWeight: categoryDiversifyWeight,
       ),
-    );
-  }
-
-  @override
-  List<NavalMoveOrder> suggestNavalMoveOrders(
-    PlayerView view,
-    Game game,
-    MapTopology topology,
-    Orders currentOrders, {
-    OrderResolutionContext? resolution,
-  }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
-    return ctx.loggedSuggest(
-      'suggestNavalMoveOrders',
-      () => suggestion.suggestNavalMoveOrders(
-        ctx.view,
-        ctx.game,
-        ctx.topology,
-        ctx.currentOrders,
-        resolution: resolution,
-      ),
-    );
-  }
-
-  @override
-  List<NavalMissionOrder> suggestNavalMissionOrders(
-    PlayerView view,
-    Game game,
-    MapTopology topology,
-    Orders currentOrders, {
-    OrderResolutionContext? resolution,
-  }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
-    return ctx.loggedSuggest(
-      'suggestNavalMissionOrders',
-      () => suggestion.suggestNavalMissionOrders(
-        ctx.view,
-        ctx.game,
-        ctx.topology,
-        ctx.currentOrders,
-        resolution: resolution,
-      ),
-    );
-  }
-
-  @override
-  List<DiplomaticOrder> suggestDiplomaticOrders(
-    PlayerView view,
-    Game game,
-    MapTopology topology,
-    Orders currentOrders, {
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
-    return ctx.loggedSuggest(
-      'suggestDiplomaticOrders',
-      () => suggestion.suggestDiplomaticOrders(
-        ctx.view,
-        ctx.game,
-        ctx.topology,
-        ctx.currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      ),
-    );
-  }
-
-  @override
-  List<DiplomaticOrder> suggestDeclareWarOrders(
-    PlayerView view,
-    Game game,
-    MapTopology topology,
-    Orders currentOrders, {
-    Map<String, TileMapResult>? tileMapByRegion,
-  }) {
-    final ctx = _ctx(view, game, topology, currentOrders);
-    return ctx.loggedSuggest(
-      'suggestDeclareWarOrders',
-      () => suggestion.suggestDeclareWarOrders(
-        ctx.view,
-        ctx.game,
-        ctx.topology,
-        ctx.currentOrders,
-        tileMapByRegion: tileMapByRegion,
-      ),
-    );
-  }
-
-  @override
-  TradeSuggestionResult suggestTradeOrders(
-    PlayerView view,
-    Game game, {
-    TradeSuggestionContext? contextOverride,
-  }) {
-    ordersLog.d(
-      'order suggestion API suggestTradeOrders player=${view.playerId}',
-    );
-    if (contextOverride != null) {
-      return TradeOrderSuggester.suggest(contextOverride);
-    }
-    final context = _defaultTradeSuggestionContext(view, game);
-    return TradeOrderSuggester.suggest(context);
-  }
-
-  TradeSuggestionContext _defaultTradeSuggestionContext(
-    PlayerView view,
-    Game game,
-  ) {
-    final player = game.playerById(view.playerId);
-    final available = <CommodityId, int>{};
-    if (player != null) {
-      for (final entry in player.stockpile.quantities.entries) {
-        if (richesCommodityIds.contains(entry.key)) continue;
-        if (entry.value <= 0) continue;
-        available[entry.key] = entry.value;
-      }
-    }
-    return tradeSuggestionContextFromGame(
-      game,
-      view.playerId,
-      availableStockpileByCommodityId: available,
     );
   }
 }
