+++ packages/colonizethis_orders/lib/src/orders/work_handlers/standard_work_handler.dart
@@ -1,193 +1,11 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
-import '../orders_logging.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
+import 'package:colonizethis_world/colonizethis_world.dart';
 
 import '../order_work_constants.dart';
 import '../build_rail_work_rules.dart';
-import '../validators/work_order_cost_calculator.dart';
+import 'standard_work_handler_config.dart';
 import 'work_order_handler.dart';
-
-class _StandardWorkTargetConfig {
-  const _StandardWorkTargetConfig({
-    required this.allowedForUnitType,
-    required this.costFn,
-    required this.totalTurnsFn,
-  });
-
-  final bool Function(String) allowedForUnitType;
-  final WorkOrderCost? Function() costFn;
-  final int Function() totalTurnsFn;
-}
-
-/// Declarative cost/turn strategy for each standard build work target (Refs
-/// #3877). Fixed-material targets share one resolver; improvement and fort
-/// retain context-sensitive cost/turn wiring.
-enum _StandardWorkTargetKind {
-  fixedMaterial,
-  improvement,
-  fort,
-}
-
-const Map<String, _StandardWorkTargetKind> _standardWorkTargetKinds = {
-  kWorkTargetBuildImprovement: _StandardWorkTargetKind.improvement,
-  kWorkTargetBuildFort: _StandardWorkTargetKind.fort,
-  kWorkTargetBuildRoad: _StandardWorkTargetKind.fixedMaterial,
-  kWorkTargetBuildPort: _StandardWorkTargetKind.fixedMaterial,
-  kWorkTargetBuildRail: _StandardWorkTargetKind.fixedMaterial,
-  kWorkTargetUpgradeTown: _StandardWorkTargetKind.fixedMaterial,
-};
-
-const _StandardWorkTargetConfig _unsupportedStandardWorkTargetConfig =
-    _StandardWorkTargetConfig(
-      allowedForUnitType: _alwaysFalseForWorkTarget,
-      costFn: _nullWorkOrderCost,
-      totalTurnsFn: _singleTurnWorkDuration,
-    );
-
-_StandardWorkTargetConfig _buildStandardWorkTargetConfig({
-  required Game game,
-  required String target,
-  required String targetTileKey,
-  required Unit unit,
-  required TileMapState tileState,
-  required Map<String, Province> provincesById,
-}) {
-  final kind = _standardWorkTargetKinds[target];
-  if (kind == null) return _unsupportedStandardWorkTargetConfig;
-  return switch (kind) {
-    _StandardWorkTargetKind.fixedMaterial => _StandardWorkTargetConfig(
-      allowedForUnitType: (t) =>
-          isWorkOrderTargetAllowedForUnitType(t, target),
-      costFn: () => workOrderMaterialCost(target),
-      totalTurnsFn: () => totalTurnsForWork(target),
-    ),
-    _StandardWorkTargetKind.improvement => _StandardWorkTargetConfig(
-      allowedForUnitType: (t) => isWorkOrderTargetAllowedForUnitType(
-        t,
-        kWorkTargetBuildImprovement,
-      ),
-      costFn: () => WorkOrderCostCalculator(
-        game,
-        playerId: unit.ownerId,
-      ).calculateCost(
-        kWorkTargetBuildImprovement,
-        targetTileKey,
-        improvementLevel: tileState.improvementLevel(targetTileKey),
-      ),
-      totalTurnsFn: () => totalTurnsForWork(
-        kWorkTargetBuildImprovement,
-        improvementLevel: tileState.improvementLevel(targetTileKey),
-      ),
-    ),
-    _StandardWorkTargetKind.fort => _fortWorkTargetConfig(
-      unit: unit,
-      provincesById: provincesById,
-    ),
-  };
-}
-
-_StandardWorkTargetConfig _fortWorkTargetConfig({
-  required Unit unit,
-  required Map<String, Province> provincesById,
-}) {
-  final fortLevel = provincesById[unit.locationProvinceId]?.fortLevel ?? 0;
-  return _StandardWorkTargetConfig(
-    allowedForUnitType: (t) =>
-        isWorkOrderTargetAllowedForUnitType(t, kWorkTargetBuildFort),
-    costFn: () =>
-        workOrderMaterialCost(kWorkTargetBuildFort, fortLevel: fortLevel),
-    totalTurnsFn: () =>
-        totalTurnsForWork(kWorkTargetBuildFort, fortLevel: fortLevel),
-  );
-}
-
-bool _alwaysFalseForWorkTarget(String _) => false;
-WorkOrderCost? _nullWorkOrderCost() => null;
-int _singleTurnWorkDuration() => 1;
-
-bool applyStandardWorkOrder({
-  required Game game,
-  required WorkOrder order,
-  required Unit unit,
-  required String targetTileKey,
-  required bool hasValidTarget,
-  required String orderTarget,
-  required TileMapState tileState,
-  required Map<String, Province> provincesById,
-  required bool Function(WorkOrderCost) canAffordMaterialCost,
-  required void Function(WorkOrderCost) deductMaterialCost,
-  required void Function(String, Unit) updateUnit,
-}) {
-  if (unit.currentWork != null || !hasValidTarget) return false;
-
-  final config = _buildStandardWorkTargetConfig(
-    game: game,
-    target: orderTarget,
-    targetTileKey: targetTileKey,
-    unit: unit,
-    tileState: tileState,
-    provincesById: provincesById,
-  );
-  if (!config.allowedForUnitType(unit.type)) return false;
-
-  final cost = config.costFn();
-  if (cost == null || !canAffordMaterialCost(cost)) return false;
-
-  deductMaterialCost(cost);
-  final totalTurns = config.totalTurnsFn();
-  ordersLog.d(
-    'work order accepted and assigned unit=${order.unitId} target=$orderTarget targetTileKey=$targetTileKey totalTurns=$totalTurns',
-  );
-  updateUnit(
-    order.unitId,
-    unit.copyWith(
-      status: UnitStatus.working,
-      tileKey: targetTileKey,
-      originTileKey: unit.originTileKey ?? unit.tileKey,
-      assignedTileKey: targetTileKey,
-      currentWork: CurrentWork(
-        workTarget: orderTarget,
-        tileKey: targetTileKey,
-        totalTurns: totalTurns,
-        remainingTurns: totalTurns,
-      ),
-    ),
-  );
-  return true;
-}
-
-bool shouldSkipBuildFortForMissingTech({
-  required Province? province,
-  required Map<String, bool>? techUnlocked,
-}) {
-  final fortLevel = province?.fortLevel ?? 0;
-  if (fortLevel == 1 && techUnlocked?[kTechIdMineEngineering] != true) {
-    ordersLog.d(
-      'build_fort skipped - Mine Engineering required for fort level 2',
-    );
-    return true;
-  }
-  if (fortLevel == 2 && techUnlocked?[kTechIdModernForts] != true) {
-    ordersLog.d('build_fort skipped - Modern Forts required for fort level 3');
-    return true;
-  }
-  return false;
-}
-
-bool shouldSkipBuildRailForInvalidTerrainOrTech({
-  required Map<String, bool>? techUnlocked,
-  required int roadLevel,
-  required TerrainType? terrain,
-}) {
-  final railReason = rejectionReasonForBuildRailOrder(
-    techUnlocked: techUnlocked,
-    roadLevel: roadLevel,
-    terrain: terrain,
-  );
-  if (railReason == null) return false;
-  ordersLog.d('build_rail skipped - $railReason');
-  return true;
-}
 
 typedef StandardBuildPreApplyGate =
     bool Function({
